requirements.txt
streamlit
ollama
edge-tts
SpeechRecognition
pyaudio
pygame
python-dotenv
openai-whisper
